+++ priv/repo/migrations/20210914115259_add_refresh_token_ttl_to_clients.exs
@@ -1,0 +1,13 @@
+defmodule Boruta.Repo.Migrations.AddRefreshTokenTtlToClients do
+  use Ecto.Migration
+
+  def change do
+    alter table(:clients) do
+      add :refresh_token_ttl, :integer, null: false, default: "2592000"
+    end
+
+    alter table(:clients) do
+      modify :refresh_token_ttl, :integer, null: false
+    end
+  end
+end