+++ priv/repo/migrations/20210926200753_add_public_revoke_to_clients.exs
@@ -1,0 +1,9 @@
+defmodule Boruta.Repo.Migrations.AddPublicRevokeToClients do
+  use Ecto.Migration
+
+  def change do
+    alter table(:oauth_clients) do
+      add :public_revoke, :boolean, null: false, default: false
+    end
+  end
+end