+++ test/boruta/oauth/integration/implicit_grant_test.exs
@@ -19,7 +19,6 @@ defmodule Boruta.OauthTest.ImplicitGrantTest do
       user = %User{}
       resource_owner = %ResourceOwner{sub: user.id, username: user.email}
       client = insert(:client, redirect_uris: ["https://redirect.uri"])
-      client_with_regex = insert(:client, redirect_uris: ["https://(.+).uri"])
       client_without_grant_type = insert(:client, supported_grant_types: [])
 
       client_with_scope =
@@ -32,7 +31,6 @@ defmodule Boruta.OauthTest.ImplicitGrantTest do
       {:ok,
        client: client,
        client_with_scope: client_with_scope,
-       client_with_regex: client_with_regex,
        client_without_grant_type: client_without_grant_type,
        resource_owner: resource_owner}
     end
@@ -343,39 +341,6 @@ defmodule Boruta.OauthTest.ImplicitGrantTest do
                "sub" => ^resource_owner_id,
                "nonce" => ^nonce
       } = claims
-    end
-
-    test "returns a token with regexes", %{
-      client_with_regex: client,
-      resource_owner: resource_owner
-    } do
-      ResourceOwners
-      |> stub(:get_by, fn _params -> {:ok, resource_owner} end)
-
-      redirect_uri = List.first(client.redirect_uris)
-
-      {:authorize_success,
-       %AuthorizeResponse{
-         type: type,
-         access_token: value,
-         expires_in: expires_in,
-         token_type: "bearer"
-       }} =
-        Oauth.authorize(
-          %Plug.Conn{
-            query_params: %{
-              "response_type" => "token",
-              "client_id" => client.id,
-              "redirect_uri" => redirect_uri
-            }
-          },
-          resource_owner,
-          ApplicationMock
-        )
-
-      assert type == :token
-      assert value
-      assert expires_in
     end
 
     test "returns a token if scope is authorized", %{