+++ test/boruta/oauth/integration/revoke_test.exs
@@ -15,6 +15,7 @@ defmodule Boruta.OauthTest.RevokeTest do
   describe "revoke request" do
     setup do
       client = insert(:client)
+      public_revoke_client = insert(:client, public_revoke: true)
       user = %User{}
       resource_owner = %ResourceOwner{sub: user.id, username: user.email}
       token = insert(:token,
@@ -25,6 +26,7 @@ defmodule Boruta.OauthTest.RevokeTest do
       )
       {:ok,
         client: client,
+        public_revoke_client: public_revoke_client,
         token: token,
         resource_owner: resource_owner
       }
@@ -33,7 +35,7 @@ defmodule Boruta.OauthTest.RevokeTest do
     test "returns an error without params" do
       assert Oauth.revoke(%Plug.Conn{body_params: %{}}, ApplicationMock) == {:revoke_error, %Error{
         error: :invalid_request,
-        error_description: "Request validation failed. Required properties client_id, client_secret, token are missing at #.",
+        error_description: "Request validation failed. Required properties client_id, token are missing at #.",
         status: :bad_request
       }}
     end
@@ -41,7 +43,7 @@ defmodule Boruta.OauthTest.RevokeTest do
     test "returns an error with invalid request" do
       assert Oauth.revoke(%Plug.Conn{body_params: %{}}, ApplicationMock) == {:revoke_error, %Error{
         error: :invalid_request,
-        error_description: "Request validation failed. Required properties client_id, client_secret, token are missing at #.",
+        error_description: "Request validation failed. Required properties client_id, token are missing at #.",
         status: :bad_request
       }}
     end
@@ -63,6 +65,7 @@ defmodule Boruta.OauthTest.RevokeTest do
       ResourceOwners
       |> stub(:get_by, fn(_params) -> {:ok, resource_owner} end)
       %{req_headers: [{"authorization", authorization_header}]} = using_basic_auth(client.id, client.secret)
+
       case Oauth.revoke(%Plug.Conn{
         body_params: %{"token" => token.value},
         req_headers: [{"authorization", authorization_header}]
@@ -71,6 +74,15 @@ defmodule Boruta.OauthTest.RevokeTest do
           assert Boruta.AccessTokensAdapter.get_by(value: token.value).revoked_at
         _ -> assert false
       end
+    end
+
+    test "revoke token if client has public revocation", %{public_revoke_client: client, token: token, resource_owner: resource_owner} do
+      ResourceOwners
+      |> stub(:get_by, fn(_params) -> {:ok, resource_owner} end)
+
+      assert {:revoke_success} = Oauth.revoke(%Plug.Conn{
+        body_params: %{"token" => token.value, "client_id" => client.id},
+      }, ApplicationMock)
     end
   end