+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+  ### Added
+    - `public_refresh_token` per client configuration allowing to refresh tokens without providing client secret
   ### Fixed
     - `boruta.gen.controllers` generated paths in umbrella apps
     - Refreshed tokens has associated access_token scope as default
+    - Requests with no client secret won't raise an error
   ### Changed
     - `invalid_client` do not return neither format, nor redirect_uri in `Boruta.Oauth.Error`
 ## [1.1.0]